+++ packages/vue-router/README.md
@@ -9,8 +9,10 @@ import BugsnagPerformance from '@bugsnag/browser-performance'
 import { VueRouterRoutingProvider } from '@bugsnag/vue-router-performance'
 import { createRouter, createWebHistory } from 'vue-router'
 
+const base = 'my-app
+
 const router = createRouter({
-  history: createWebHistory('vue-router'),
+  history: createWebHistory(base),
   routes: [
     {
       path: '/',
@@ -20,16 +22,15 @@ const router = createRouter({
     {
       path: '/contacts/:contactId',
       name: 'contact',
-      component: () => import('../views/ContactView.vue')
+      component: () => import('./views/ContactView.vue')
     }
   ]
 })
 
-const basename = '/vue-router'
 
 BugsnagPerformance.start({
   apiKey,
-  routingProvider: new VueRouterRoutingProvider(router, basename)
+  routingProvider: new VueRouterRoutingProvider(router, base)
 })
 
 const app = createApp(App)
