+++ test/browser/features/fixtures/packages/vue-router/src/main.js
@@ -1,4 +1,6 @@
 import { createApp } from 'vue'
+import { createRouter, createWebHistory } from 'vue-router'
+import HomeView from './views/HomeView.vue'
 import App from './App.vue'
 import router from './router'
 import BugsnagPerformance from '@bugsnag/browser-performance'
@@ -8,7 +10,24 @@ const parameters = new URLSearchParams(window.location.search)
 const apiKey = parameters.get('api_key')
 const endpoint = parameters.get('endpoint')
 
-const basename = '/vue-router'
+const base = '/vue-router'
+
+const router = createRouter({
+  history: createWebHistory(base),
+  routes: [
+    {
+      path: '/',
+      name: 'home',
+      component: HomeView
+    },
+    {
+      path: '/contacts/:contactId',
+      name: 'contact',
+      component: () => import('./views/ContactView.vue')
+    }
+  ]
+})
+
 
 BugsnagPerformance.start({
     apiKey,
@@ -16,7 +35,7 @@ BugsnagPerformance.start({
     maximumBatchSize: 13,
     batchInactivityTimeoutMs: 5000,
     autoInstrumentNetworkRequests: false,
-    routingProvider: new VueRouterRoutingProvider(router, basename)
+    routingProvider: new VueRouterRoutingProvider(router, base)
 })
 
 const app = createApp(App)
